requirements.txt
google-genai
python-dotenv
sentence-transformers
numpy
PyPDF2
groq
schedule 
streamlit
schedule
requests 
pandas